+++ spec/validb/console_logger_spec.rb
@@ -6,7 +6,7 @@ describe Validb::ConsoleLogger do
       hash = { model: Blog, id: 1, error_messages: "Error message" }
 
       console_logger = Validb::ConsoleLogger.new
-      $stdout.should_receive(:puts).with("Blog:1 - Error message")
+      $stdout.should_receive(:puts).with("\nBlog:1 - Error message")
       console_logger.out(hash)
     end
   end